Attempting to fix import problems by fixing __init__'s; also some renaming
Note updated main slightly as well
