+++ brains/__init__.py
@@ -1,0 +1,2 @@
+from . import data
+from . import viz
